# feat(interpreter): Unify instruction fn signature