--- a/bins/revme/src/statetest/runner.rs
+++ b/bins/revme/src/statetest/runner.rs
@@ -277,8 +277,7 @@ pub fn execute_test_suit(path: &Path, elapsed: &Arc<Mutex<Duration>>) -> Result<
                     let db = evm.db().unwrap();
                     println!("{path:?} UNIT_TEST:{name}\n");
                     println!(
-                        "failed reason: {:?} {:?} UNIT_TEST:{}\n gas:{:?} ({:?} refunded)",
-                        exit_reason, path, name, gas_used, gas_refunded,
+                        "failed reason: {exit_reason:?} {path:?} UNIT_TEST:{name}\n gas:{gas_used:?} ({gas_refunded:?} refunded)",
                     );
                     println!("\nApplied state:{db:?}\n");
                     println!("\nStateroot: {state_root:?}\n");