All opcode instructions are now functions  with same signature:
`pub fn jumpi(interpreter: &mut Interpreter, _host: &mut dyn Host)` 

This is revm internal change and it does not affect external API.

This allows the compiler to optimize calls in the interpreter hot loop and it gives around ~10% faster `snailtracer` execution on i7. I am interested to see if the same boost can be found on macs with `M` processors.

This is one of steps needed to define multiple dispatch strategies related to interpreter loop. As instructions now have same signature they can be put in an array, this can have potentially new ways to use revm as it is easier to just switch function pointer to something else (custom function to sload/sstore?)

`Spec` was simplified and `is_static` is moved as variable to interpreter (Previously it was inside Spec)

There was one regression with ruint in casting U256 to saturated usize [here](https://github.com/bluealloy/revm/blob/acdbaaca5d110c32bdd1f7ecd7974a306a40a5e2/crates/revm/src/instructions/macros.rs#L236-L253). It was faster to just check limbs.

`microbench` added to `revm-tests`