@rakita 2022-12-04
new run (with microbench)
```
Snailtracer benchmark (3.4s) ...  41_297_247.647 ns/iter (0.999 R²)
```

old run:
```
elapsed: 47.478018ms
0: 47.639335ms
1: 47.21674ms
2: 47.112436ms
3: 46.998897ms
4: 46.768525ms
5: 46.722415ms
6: 46.711912ms
7: 46.705151ms
8: 46.658515ms
9: 46.649658ms
10: 46.642895ms
11: 46.631915ms
12: 46.622408ms
13: 46.618776ms
14: 46.617359ms
15: 46.612672ms
16: 46.605036ms
17: 46.58608ms
18: 46.543365ms
19: 46.531848ms
20: 46.524966ms
21: 46.523186ms
22: 46.521023ms
23: 46.514446ms
24: 46.496036ms
end!
```

There is not a lot that can be seen from flamegraph, but there they are:

flamegraph new:
![revm_unification](https://user-images.githubusercontent.com/13179220/205494903-8d3c7f27-1157-4e55-ab6d-824b0cca11ea.svg)


flamegraph old:
![revm_main_before_unitication](https://user-images.githubusercontent.com/13179220/205494923-556f43ce-f858-478e-8cd9-bbd5ad056a62.svg)