@gakonst 2022-12-09
@ziyadedher can you rerun https://github.com/ziyadedher/evm-bench with latest main? ^^