@rakita 2022-12-09
I have one more round of small optimizations, but this should give a good boost in performance.