@ziyadedher 2022-12-17
I also see ~10% bumps on snailtracer! But interestingly, I don't see any noticeable differences on other benches, see https://github.com/ziyadedher/evm-bench/commit/aad8de9d7de5a301f834c45a2983a5d8646ea1c5

This might make sense since the optimization is primarily in dispatches and whatnot it seems. Just wanted to float the results!